Refactor: 이미지 서비스 코드 사용 리팩토링
- 도메인 서비스에서 파사드 메소드만 사용하도록 리팩토링
- Guide, GuideBook 서비스는 삭제 이미지 리스트 추가 후, 반영 필요.
