+++ src/main/java/trendravel/photoravel_be/domain/guidebook/service/GuidebookService.java
@@ -4,12 +4,12 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+import trendravel.photoravel_be.commom.image.service.ImageService;
 import trendravel.photoravel_be.domain.guidebook.dto.request.GuidebookRequestDto;
 import trendravel.photoravel_be.domain.guidebook.dto.response.GuidebookResponseDto;
 import trendravel.photoravel_be.db.guidebook.Guidebook;
 import trendravel.photoravel_be.db.enums.Region;
 import trendravel.photoravel_be.db.respository.guidebook.GuidebookRepository;
-import trendravel.photoravel_be.commom.service.ImageService;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -33,11 +33,11 @@ public class GuidebookService {
                 .title(guidebookRequestDto.getTitle())
                 .content(guidebookRequestDto.getContent())
                 .region(guidebookRequestDto.getRegion())
-                .images(imageService.uploadImages(images))
+                .images(imageService.uploadImageFacade(images))
                 .views(0)
                 .build());
-        
-        
+
+        imageService.uploadImages(images);
         return GuidebookResponseDto.builder()
                 .id(guidebook.getId())
                 .userId(guidebook.getUserId())
@@ -137,8 +137,12 @@ public class GuidebookService {
         }
         
         Guidebook guidebook = guidebookOpt.get();
-        guidebook.updateGuidebook(guidebookRequestDto, imageService.uploadImages(images));
-        
+
+        /**
+         * 삭제 이미지 리스트 필요, 수정 필요.
+         */
+//        guidebook.updateGuidebook(guidebookRequestDto, imageService.updateImageFacade(images));
+        imageService.uploadImages(images);
         return GuidebookResponseDto.builder()
                 .id(guidebook.getId())
                 .userId(guidebook.getUserId())
