+++ src/main/java/trendravel/photoravel_be/domain/location/service/LocationService.java
@@ -2,6 +2,7 @@ package trendravel.photoravel_be.domain.location.service;
 
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.jetbrains.annotations.NotNull;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.GeometryFactory;
@@ -32,6 +33,7 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class LocationService {
 
     private final LocationRepository locationRepository;
@@ -40,9 +42,11 @@ public class LocationService {
     @Transactional
     public LocationResponseDto createLocation(
             LocationRequestDto locationRequestDto, List<MultipartFile> images) {
+
         Location location = Location.builder()
                 .description(locationRequestDto.getDescription())
                 .name(locationRequestDto.getName())
+                .images(imageService.uploadImageFacade(images))
                 .latitude(locationRequestDto.getLatitude())
                 .longitude(locationRequestDto.getLongitude())
                 .address(locationRequestDto.getAddress())
@@ -53,7 +57,6 @@ public class LocationService {
                 .build();
         location.getPoint().setSRID(4326);
         locationRepository.save(location);
-        location.createLocationImage(imageService.uploadImages(images));
 
 
         return LocationResponseDto
@@ -173,16 +176,16 @@ public class LocationService {
 
     @Transactional
     public LocationResponseDto updateLocation(
-            LocationUpdateImagesDto locationRequestDto, List<MultipartFile> images) {
+            LocationUpdateImagesDto locationRequestDto,
+            List<MultipartFile> images) {
 
         Location location = locationRepository.findById(
                 locationRequestDto.getLocationId())
                 .orElseThrow(() -> new ApiException(LocationErrorCode.LOCATION_NOT_FOUND));
-
-
+        log.info("이미지 저장 전");
         location.updateLocation(locationRequestDto,
-                imageService.updateImages(images, locationRequestDto.getDeleteImages()));
-
+                imageService.updateImageFacade(images, locationRequestDto.getDeleteImages()));
+        log.info("이미지 저장 후");
 
         return LocationResponseDto
                 .builder()
@@ -227,9 +230,7 @@ public class LocationService {
         Location findLocation = locationRepository.findById(id)
                 .orElseThrow(() -> new ApiException(LocationErrorCode.LOCATION_NOT_FOUND));
         locationRepository.deleteById(findLocation.getId());
-        imageService.deleteAllImages(findLocation.getImages());
+        imageService.deleteAllImagesFacade(findLocation.getImages());
     }
-
-
 
 }
